alterações variável Aula5.java

+++ Aula5-AC1/src/Aula5.java
@@ -4,7 +4,7 @@ public class Aula5 {
 
     public static void main(String[] args) {
         //inicialização de variáveis
-        float mediaFinal, ac1, ac2, ag, af = 0.0;
+        double mediaFinal, ac1, ac2, ag, af = 0.0;
 
 
         //Criação do scanner para a execução de todos os exercícios
@@ -12,16 +12,16 @@ public class Aula5 {
 
         //Recebe as notas da AC1, AC2, AG E AF
         System.out.print("Digite a nota da AC1: ");
-        ac1 = scanner.nextFloat();
+        ac1 = scanner.nextDouble();
 
         System.out.print("Digite a nota da AC2: ");
-        ac2 = scanner.nextFloat();
+        ac2 = scanner.nextDouble();
 
         System.out.print("Digite a nota da AG: ");
-        ag = scanner.nextFloat();
+        ag = scanner.nextDouble();
 
         System.out.print("Digite a nota da AF: ");
-        af = scanner.nextFloat();
+        af = scanner.nextDouble();
         
         //Execução do método para calcular a média final do aluno
         mediaFinal = calculoMedia(ac1, ac2, ag, af);
@@ -36,7 +36,7 @@ public class Aula5 {
     
 
     //Criação do método para calcular a média
-    public static float calculoMedia(float ac1, float ac2, float ag, float af) {
+    public static double calculoMedia(double ac1, double ac2, double ag, double af) {
         return  (ac1 * 0.15) + (ac2 * 0.30) + (ag * 0.10) + (af * 0.45);
         
     }
